requirements.txt
black==25.1.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flake8==7.2.0
isort==6.0.1
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
